+++ test/cmd/colour-manipulation.js
@@ -28,6 +28,7 @@
 // Test suite.
 describe('Colour Manipulation', () => {
   require('./colour-manipulation/greyscale')
+  require('./colour-manipulation/pipeline-colourspace')
   require('./colour-manipulation/tint')
   require('./colour-manipulation/tocolourspace')
 })